HW1/requirements.txt
imageio==2.21.2
matplotlib==3.7.5
numpy==1.24.4
Pillow==10.4.0
pandas==2.0.3
scipy==1.10.1
torch==2.4.0
torchvision==0.19.0
tensorboard==2.14.0
timm==1.0.19
gdown